% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_ui.R
\name{add_class}
\alias{add_class}
\title{TSRtracker_utils_ui}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{Tag}

\item{class}{CSS class to add}
}
\value{
appends the css class
}
\description{
Add CSS class
}
