% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{factop}
\alias{factop}
\title{adjust the opacity of the polygon in leaflet}
\usage{
factop(x)
}
\value{
opacity of polygons
}
\description{
adjust the opacity of the polygon in leaflet
}
