% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{get_word_docx}
\alias{get_word_docx}
\title{convert word docx to data.frame}
\usage{
get_word_docx(word_doc)
}
\value{
list of data.frames containing workplan information
}
\description{
convert word docx to data.frame
}
