% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{getInformationSchemaTables}
\alias{getInformationSchemaTables}
\title{Get schema tables from information_schema database}
\usage{
getInformationSchemaTables(schema)
}
\arguments{
\item{schema}{}
}
\value{
the table of schemas
}
\description{
Get schema tables from information_schema database
}
