% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{getTableQuery}
\alias{getTableQuery}
\title{Get the results of an SQL query from the database}
\usage{
getTableQuery(sql, params = list(), conn = NULL)
}
\arguments{
\item{sql}{The SQL query}

\item{params}{The SQL query parameters}

\item{conn}{Optional database connection which can be reused}
}
\value{
result set as a data.frame
}
\description{
Get the results of an SQL query from the database
}
