% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{getWordDocx}
\alias{getWordDocx}
\title{convert word docx to data.frame}
\usage{
getWordDocx(word_doc)
}
\value{
list of data.frames containing workplan information
}
\description{
convert word docx to data.frame
}
