% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{setScheduleQuery}
\alias{setScheduleQuery}
\title{Set the results of a workplan upload to a PostgreSQL database}
\usage{
setScheduleQuery(wrkpln, reportList)
}
\arguments{
\item{sql}{The SQL query}
}
\value{
void
}
\description{
Set the results of a workplan upload to a PostgreSQL database
}
