% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_TSRtracker_utils_server.R
\name{theme_gantt}
\alias{theme_gantt}
\title{Styling for the gantt chart}
\usage{
theme_gantt(base_size = 11, base_family = "Source Sans Pro Light")
}
\value{
theme for the gantt
}
\description{
Styling for the gantt chart
}
